requirements.txt
astroid==1.6.6
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.4.4
beautifulsoup4==4.8.2
behave==1.2.6
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
contextlib2==0.5.5
entrypoints==0.3
flake8
Flask==1.1.1
Flask-Cors==3.0.8
flask-swagger-ui==3.20.9
get==2019.4.13
google==2.0.3
google-api-core==1.15.0
google-auth==1.10.0
google-cloud==0.34.0
google-cloud-core==1.1.0
google-cloud-firestore==1.6.1
google-cloud-logging
google-cloud-pubsub==1.1.0
googleapis-common-protos==1.6.0
grpc-google-iam-v1==0.12.3
grpcio==1.26.0
idna==2.8
importlib-metadata==1.4.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
linecache2==1.0.0
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.1.0
packaging==20.0
parse==1.14.0
parse-type==0.5.2
pika==1.1.0
pluggy==0.13.1
post==2019.4.13
protobuf==3.11.2
public==2019.4.13
py==1.8.1
pyaml==19.12.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.6
pytest==4.6.9
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
requests-mock==1.7.0
rsa==4.0
schema==0.7.1
six==1.13.0
soupsieve==1.9.5
structlog==19.2.0
urllib3==1.25.7
virtualenv
wcwidth==0.1.8
Werkzeug==0.16.0
wrapt==1.11.2
zipp==1.0.0
